+++ R/user-activity.R
@@ -10,7 +10,7 @@
 #' 
 #' @details 
 #' 
-#' The
+#' The User Activity API lets you query an individual user's movement through your website, by sending in the individual `clientId` or `userId`.
 #' 
 #' @export
 #' 
@@ -23,8 +23,8 @@
 #' 
 #' # access data for individual users
 #' uar <- ga_clientid_activity(c("1106980347.1461227730", "476443645.1541099566"),
-#'                      viewId = 81416156, 
-#'                      date_range = c("2019-01-01","2019-02-01"))
+#'                          viewId = 81416156, 
+#'                          date_range = c("2019-01-01","2019-02-01"))
 #' 
 #' # access the data.frames returned:
 #' 
@@ -33,7 +33,14 @@
 #' 
 #' # the hit level activity for the users passed in
 #' uar$hits
-#'                      
+#'
+#' # filter the response to only include certain activity types, such as goals:
+#' 
+#' only_goals <- ga_clientid_activity(c("1106980347.1461227730", 
+#'                                      "476443645.1541099566"),
+#'                      viewId = 81416156, 
+#'                      date_range = c("2019-01-01","2019-02-01"),
+#'                      activity_types = "GOAL")                      
 #' 
 #' 
 #' }