+++ R/options.R
@@ -25,55 +25,6 @@
     googleAuthR.batch_endpoint = "https://www.googleapis.com/batch/analytics/v3",
     googleAuthR.tryAttempts = 1)
   
-  if(Sys.getenv("GAR_CLIENTID") != ""){
-    Sys.setenv("GA_CLIENT_ID" = Sys.getenv("GAR_CLIENTID"))
-  } 
-  
-  if(Sys.getenv("GAR_CLIENT_SECRET") != ""){
-    Sys.setenv("GA_CLIENT_SECRET" = Sys.getenv("GAR_CLIENT_SECRET"))
-  }
-  
-  if(Sys.getenv("GAR_WEB_CLIENTID") != ""){
-    Sys.setenv("GA_WEB_CLIENT_ID" = Sys.getenv("GAR_WEB_CLIENTID"))
-  } 
-  
-  if(Sys.getenv("GAR_WEB_CLIENT_SECRET") != ""){
-    Sys.setenv("GA_WEB_CLIENT_SECRET" = Sys.getenv("GAR_WEB_CLIENT_SECRET"))
-  }
-  
-  if(Sys.getenv("GAR_SCOPES") != ""){
-    options(googleAuthR.scopes.selected = Sys.getenv("GAR_SCOPES"))
-  }
-  
-  if(Sys.getenv("GA_CLIENT_ID") != ""){
-    options(googleAuthR.client_id = Sys.getenv("GA_CLIENT_ID"))
-  }
-  
-  if(Sys.getenv("GA_CLIENT_SECRET") != ""){
-    options(googleAuthR.client_secret = Sys.getenv("GA_CLIENT_SECRET"))
-  }
-  
-  
-  if(Sys.getenv("GA_WEB_CLIENT_ID") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_ID"))
-  }
-  
-  if(Sys.getenv("GA_WEB_CLIENT_SECRET") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_SECRET"))
-  }
-  
-  if(Sys.getenv("GAR_CLIENT_JSON") != ""){
-    packageStartupMessage("Setting your own client.id")
-    googleAuthR::gar_set_client(json = Sys.getenv("GAR_CLIENT_JSON"))
-  }
-  
-  if(Sys.getenv("GAR_CLIENT_WEB_JSON") != ""){
-    packageStartupMessage("Setting your own web client.id")
-    googleAuthR::gar_set_client(web_json = Sys.getenv("GAR_CLIENT_WEB_JSON"))
-  }
-  
-
-  
   f <- function(req){
     
     stuff <- tryCatch(req$content$reports, error = function(x) NULL)
@@ -102,20 +53,8 @@
 
 .onAttach <- function(libname, pkgname){
   
-  if(Sys.getenv("GA_CLIENT_ID") != ""){
-    options(googleAuthR.client_id = Sys.getenv("GA_CLIENT_ID"))
-  }
-  
-  if(Sys.getenv("GA_CLIENT_SECRET") != ""){
-    options(googleAuthR.client_secret = Sys.getenv("GA_CLIENT_SECRET"))
-  }
-  
-  if(Sys.getenv("GA_WEB_CLIENT_ID") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_ID"))
-  }
-  
-  if(Sys.getenv("GA_WEB_CLIENT_SECRET") != ""){
-    options(googleAuthR.webapp.client_id = Sys.getenv("GA_WEB_CLIENT_SECRET"))
+  if(Sys.getenv("GAR_CLIENT_JSON") != ""){
+    googleAuthR::gar_set_client(json = Sys.getenv("GAR_CLIENT_JSON"))
   }
   
   needed <- c("https://www.googleapis.com/auth/analytics", 
@@ -126,6 +65,7 @@
   
   googleAuthR::gar_attach_auto_auth(needed, 
                                     environment_var = "GARGLE_EMAIL")
+  # for json files
   googleAuthR::gar_attach_auto_auth(needed, 
                                     environment_var = "GA_AUTH_FILE")